moved and refactored all code for the navbar
