+++ client/src/components/Nav/UserMenu.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import Nav from "react-bootstrap/Nav";
 import { pages } from "../../router/routes";
 import { NavLink } from "react-router-dom";
